Initial commit of convenience macro wrapper to create a rust init function for actors
