+++ actor-core/rust/wasmcloud-actor-core/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "wasmcloud-actor-core"
 description = "wasmCloud Core Actor Interface"
-version = "0.2.1"
-authors = ["wasmCloud Team"]
+version = "0.2.2"
+authors = ["wasmcloud Team"]
 edition = "2018"
 license = "Apache-2.0"
 documentation = "https://docs.rs/wasmcloud-actor-core"
@@ -11,9 +11,10 @@ keywords = ["webassembly", "wasm", "wasmcloud", "actor"]
 categories = ["wasm", "api-bindings"]
 
 [features]
-guest = ["wapc-guest"]
+guest = ["wapc-guest", "wasmcloud-actor-core-derive"]
 
 [dependencies]
+wasmcloud-actor-core-derive = { path = "../wasmcloud-actor-core-derive", optional = true }
 wapc-guest = { version = "0.4.0", optional = true }
 serde = { version = "1.0.123" , features = ["derive"] }
 serde_json = "1.0.62"
